# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "E-commerce API  ",
        "description": "API para la gestión de productos en un sistema de comercio electrónico.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Obtener lista de productos",
                "operationId": "c1ff6f862214e1896c59cfe0491ce0e8",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número de página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1,
                            "minimum": 1
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista paginada de productos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductListResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token ausente o inválido"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Crear un producto",
                "operationId": "399770d33fa4672f1216707d0e236763",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Producto creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token ausente o inválido"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/products/{product}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Obtener un producto",
                "operationId": "960e17d05ed63211b20fb4b8505f1263",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Producto obtenido exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token ausente o inválido"
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Actualizar un producto",
                "operationId": "d8f8f68cbee2f6a5c291d7940a3dc4ed",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Producto actualizado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token ausente o inválido"
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Eliminar un producto",
                "operationId": "13d9c94a21e9746d2a61b6a3247b7693",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Producto eliminado correctamente"
                    },
                    "401": {
                        "description": "Token ausente o inválido"
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Products"
                ],
                "summary": "Actualizar parcialmente un producto",
                "operationId": "8138abd5af04eadc1998c013d0a2f099",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Producto actualizado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token ausente o inválido"
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "ApiSuccessResponse": {
                "required": [
                    "success",
                    "message"
                ],
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Operación exitosa."
                    }
                },
                "type": "object"
            },
...
            "PaginationMeta": {
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "from": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "last_page": {
                        "type": "integer",
                        "example": 5
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 10
                    },
                    "to": {
                        "type": "integer",
                        "example": 10,
                        "nullable": true
                    },
                    "total": {
                        "type": "integer",
                        "example": 50
                    }
                },
                "type": "object"
            },
            "PaginationLinks": {
                "properties": {
                    "first": {
                        "type": "string",
                        "format": "uri",
                        "nullable": true
                    },
                    "last": {
                        "type": "string",
                        "format": "uri",
                        "nullable": true
                    },
                    "previous": {
                        "type": "string",
                        "format": "uri",
                        "nullable": true
                    },
                    "next": {
                        "type": "string",
                        "format": "uri",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "ProductResponse": {
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/ApiSuccessResponse"
                    },
                    {
                        "properties": {
                            "data": {
                                "$ref": "#/components/schemas/Product"
                            }
                        },
                        "type": "object"
                    }
                ]
            },
            "ProductListResponse": {
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/ApiSuccessResponse"
                    },
                    {
                        "properties": {
                            "data": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            },
                            "meta": {
                                "$ref": "#/components/schemas/PaginationMeta"
                            },
                            "links": {
                                "$ref": "#/components/schemas/PaginationLinks"
                            }
                        },
                        "type": "object"
                    }
                ]
            },
            "Product": {
                "title": "Product",
                "description": "Producto",
                "required": [
                    "id",
                    "name",
                    "description",
                    "price",
                    "stock"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Camiseta"
                    },
                    "description": {
                        "type": "string",
                        "example": "Camiseta de algodón de alta calidad"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 19.99
                    },
                    "stock": {
                        "type": "integer",
                        "example": 100
                    }
                },
                "type": "object"
            },
            "ProductInput": {
                "title": "ProductInput",
                "description": "Datos necesarios para crear o actualizar un producto",
                "required": [
                    "name",
                    "description",
                    "price",
                    "stock"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Camiseta",
                        "maxLength": 255
                    },
                    "description": {
                        "type": "string",
                        "example": "Camiseta de algodón de alta calidad"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 19.99,
                        "minimum": 0
                    },
                    "stock": {
                        "type": "integer",
                        "example": 100,
                        "minimum": 0
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token JWT en formato Bearer",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}